Update spec coverage for list-loop helpers in the Health Care Application
